Change to set artwork bitmap into notification metadata
