+++ app/src/main/java/com/geckour/q/util/Media.kt
@@ -15,7 +15,6 @@ import android.provider.MediaStore
 import android.support.v4.media.MediaMetadataCompat
 import android.support.v4.media.session.MediaSessionCompat
 import androidx.core.app.NotificationCompat
-import androidx.preference.PreferenceManager
 import com.bumptech.glide.Glide
 import com.bumptech.glide.RequestBuilder
 import com.bumptech.glide.load.engine.DiskCacheStrategy
@@ -23,7 +22,6 @@ import com.geckour.q.App
 import com.geckour.q.R
 import com.geckour.q.data.db.DB
 import com.geckour.q.data.db.model.Artist
-import com.geckour.q.data.db.model.Bool
 import com.geckour.q.data.db.model.JoinedAlbum
 import com.geckour.q.data.db.model.JoinedTrack
 import com.geckour.q.domain.model.Genre
@@ -339,23 +337,19 @@ suspend fun Song.getMediaMetadata(context: Context): MediaMetadataCompat {
         .putString(MediaMetadataCompat.METADATA_KEY_TITLE, title)
         .putString(MediaMetadataCompat.METADATA_KEY_ARTIST, artist.title)
         .putString(MediaMetadataCompat.METADATA_KEY_ALBUM, album.title)
-        .putString(MediaMetadataCompat.METADATA_KEY_ALBUM_ART_URI, uriString)
         .putString(MediaMetadataCompat.METADATA_KEY_COMPOSER, composer)
         .apply {
-            if (uriString != null && PreferenceManager.getDefaultSharedPreferences(context).showArtworkOnLockScreen) {
-                val bitmap = try {
-                    withContext(Dispatchers.IO) {
+            uriString?.let {
+                val bitmap = withContext(Dispatchers.IO) {
+                    catchAsNull {
                         Glide.with(context)
                             .asDrawable()
-                            .load(uriString)
+                            .load(it)
                             .applyDefaultSettings()
                             .submit()
                             .get()
                             .bitmap()
                     }
-                } catch (t: Throwable) {
-                    Timber.e(t)
-                    null
                 }
                 putBitmap(MediaMetadataCompat.METADATA_KEY_ALBUM_ART, bitmap)
             }
